collected the functionality in one meta class
