+++ galmock/__init__.py
@@ -1,2 +1,2 @@
 from .core.version import __version__
-from .core.datastore import DataStore
+from .galmock import GalaxyMock
